requirements.txt
# Core dependencies
flask>=2.0.0
requests>=2.25.0
yt-dlp>=2023.0.0

# Competitor recon (scraping + transcription)
# instaloader>=4.10        # Instagram scraping — pip3 install instaloader
openai-whisper>=20231117   # Local transcription — requires ffmpeg

# PDF generation
reportlab>=4.0.0

# YouTube Analytics API (OAuth)
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0

# Environment variable loading
python-dotenv>=1.0.0
